Refactor filtering search
Previously there was a lot state setting for things that didn't truly need it and there was unnecessary complexity regarding with the data arrays. Also the search code wasn't the cleanest and ultimately all this lead to the search being a bit buggy.

In this commit the array of posts is stored in a variable and the current list is set to state 'data' object. We use the same data object for displaying search results as well. Also the search bar has now internal state which further reduces the need to keep the search query in the parents state.

At the same go, rearrange the code a bit